# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Discord Bot JDA API
  description: Complete API for interacting with a Java Discord bot using JDA
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /guilds/{guildId}:
    get:
      summary: Get guild information
      description: Returns detailed information about a specific guild
      operationId: getGuild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed guild information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
        '404':
          description: Guild not found
...
  /users/{userId}:
    get:
      summary: Get user information
      description: Returns detailed information about a specific user
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /users/{userId}/dm:
    get:
      summary: Get DM channel
      description: Returns the DM channel for a specific user
      operationId: getUserDmChannel
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The DM channel for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          description: User not found

    post:
      summary: Create DM channel
      description: Creates a DM channel with a specific user
      operationId: createUserDmChannel
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: DM channel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          description: User not found

  # Events endpoints
...
components:
  schemas:
...
    Guild:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Discord ID of the guild
        name:
          type: string
          description: Name of the guild
        icon:
          type: string
          format: uri
          description: URL of the guild icon
        iconHash:
          type: string
          description: Hash of the guild icon
        splash:
          type: string
          description: Hash of the guild splash image
        discoverySplash:
          type: string
          description: Hash of the guild discovery splash image
        owner:
          type: boolean
          description: Whether the bot is the owner of the guild
        ownerId:
          type: string
          description: ID of the guild owner
        permissions:
          type: string
          description: Permissions for the bot in the guild (as a bit set)
        region:
          type: string
          description: Voice region of the guild
        afkChannelId:
          type: string
          description: ID of the AFK channel
        afkTimeout:
          type: integer
          description: AFK timeout in seconds
        widgetEnabled:
          type: boolean
          description: Whether the widget is enabled
        widgetChannelId:
          type: string
          description: Channel ID for the guild widget
        verificationLevel:
          type: integer
          description: Verification level required for the guild
        defaultMessageNotifications:
          type: integer
          description: Default message notification level
        explicitContentFilter:
          type: integer
          description: Explicit content filter level
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          description: Roles in the guild
        emojis:
          type: array
          items:
            $ref: '#/components/schemas/Emoji'
          description: Custom emojis in the guild
        features:
          type: array
          items:
            type: string
          description: Enabled guild features
        mfaLevel:
          type: integer
          description: Required MFA level for the guild
        applicationId:
          type: string
          description: Application ID of the guild creator if bot-created
        systemChannelId:
          type: string
          description: ID of the system channel
        systemChannelFlags:
          type: integer
          description: System channel flags
        rulesChannelId:
          type: string
          description: ID of the rules channel
        joinedAt:
          type: string
          format: date-time
          description: When the bot joined the guild
        large:
          type: boolean
          description: Whether this is considered a large guild
        unavailable:
          type: boolean
          description: Whether the guild is unavailable
        memberCount:
          type: integer
          description: Total number of members in the guild
        voiceStates:
          type: array
          items:
            $ref: '#/components/schemas/VoiceState'
          description: Voice states of members in the guild
        members:
          type: array
          items:
            $ref: '#/components/schemas/GuildMember'
          description: Members in the guild
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
          description: Channels in the guild
        threads:
          type: array
          items:
            $ref: '#/components/schemas/ThreadChannel'
          description: Threads in the guild
        presences:
          type: array
          items:
            $ref: '#/components/schemas/Presence'
          description: Presences of the members in the guild
        maxPresences:
          type: integer
          description: Maximum number of presences for the guild
        maxMembers:
          type: integer
          description: Maximum number of members for the guild
        vanityUrlCode:
          type: string
          description: Vanity URL code for the guild
        description:
          type: string
          description: Description of the guild
        banner:
          type: string
          description: Banner hash
        premiumTier:
          type: integer
          description: Premium tier
        premiumSubscriptionCount:
          type: integer
          description: Number of boosts this guild currently has
        preferredLocale:
          type: string
          description: Preferred locale
        publicUpdatesChannelId:
          type: string
          description: ID of the public updates channel
        maxVideoChannelUsers:
          type: integer
          description: Maximum amount of users in a video channel
        approximateMemberCount:
          type: integer
          description: Approximate number of members in this guild
        approximatePresenceCount:
          type: integer
          description: Approximate number of presences in this guild
        welcomeScreen:
          type: object
          description: Welcome screen object
        nsfwLevel:
          type: integer
          description: NSFW level
        stickers:
          type: array
          items:
            $ref: '#/components/schemas/Sticker'
          description: Custom stickers in the guild
...
    Channel:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          description: Discord ID of the channel
        type:
          type: integer
          description: Type of the channel
          enum: [0, 1, 2, 3, 4, 5, 6, 10, 11, 12, 13, 14, 15]
        guildId:
          type: string
          description: ID of the guild this channel belongs to
        position:
          type: integer
          description: Position of the channel
        permissionOverwrites:
          type: array
          items:
            $ref: '#/components/schemas/PermissionOverwrite'
          description: Permission overwrites for the channel
        name:
          type: string
          description: Name of the channel
        topic:
          type: string
          description: Topic of the channel
        nsfw:
          type: boolean
          description: Whether the channel is NSFW
        lastMessageId:
          type: string
          description: ID of the last message sent in this channel
        bitrate:
          type: integer
          description: Bitrate of a voice channel
        userLimit:
          type: integer
          description: User limit of a voice channel
        rateLimitPerUser:
          type: integer
          description: Slow mode duration in seconds
        recipients:
          type: array
          items:
            $ref: '#/components/schemas/User'
          description: Recipients of a DM
        icon:
          type: string
          description: Icon hash
        ownerId:
          type: string
          description: ID of the DM creator
        applicationId:
          type: string
          description: ID of the application that created this channel
        parentId:
          type: string
          description: ID of the parent category for this channel
        lastPinTimestamp:
          type: string
          format: date-time
          description: When the last pinned message was pinned
        rtcRegion:
          type: string
          description: Voice region of the channel
        videoQualityMode:
          type: integer
          description: Video quality mode of the channel
        messageCount:
          type: integer
          description: Number of messages in a thread
        memberCount:
          type: integer
          description: Number of members in a thread
        threadMetadata:
          $ref: '#/components/schemas/ThreadMetadata'
          description: Thread-specific fields
        member:
          $ref: '#/components/schemas/ThreadMember'
          description: Thread member for the current user
        defaultAutoArchiveDuration:
          type: integer
          description: Default auto archive duration for threads
        permissions:
          type: string
          description: Computed permissions for the invoking user
        flags:
          type: integer
          description: Channel flags
        totalMessageSent:
          type: integer
          description: Number of messages ever sent in a thread
        availableTags:
          type: array
          items:
            $ref: '#/components/schemas/ForumTag'
          description: Available tags for a forum channel
        appliedTags:
          type: array
          items:
            type: string
          description: IDs of tags applied to a forum thread
        defaultReactionEmoji:
          $ref: '#/components/schemas/DefaultReaction'
          description: Default reaction for a forum channel
        defaultThreadRateLimitPerUser:
          type: integer
          description: Default slow mode for forum threads
        defaultSortOrder:
          type: integer
          description: Default sort order for forum posts
        defaultForumLayout:
          type: integer
          description: Default forum layout view
...
    User:
      type: object
      required:
        - id
        - username
        - discriminator
      properties:
        id:
          type: string
          description: Discord ID of the user
        username:
          type: string
          description: Username of the user
        discriminator:
          type: string
          description: 4-digit discord tag
        global_name:
          type: string
          description: Global display name of the user
        avatar:
          type: string
          description: Avatar hash
        bot:
          type: boolean
          description: Whether the user is a bot
        system:
          type: boolean
          description: Whether the user is an official Discord system user
        mfaEnabled:
          type: boolean
          description: Whether the user has two-factor authentication enabled
        banner:
          type: string
          description: User banner hash
        accentColor:
          type: integer
          description: User banner color as an integer representation of a hex color
        locale:
          type: string
          description: User's chosen language option
        verified:
          type: boolean
          description: Whether the email on this account has been verified
        email:
          type: string
          format: email
          description: User's email
        flags:
          type: integer
          description: User flags
        premiumType:
          type: integer
          description: Type of Nitro subscription
        publicFlags:
          type: integer
          description: Public flags on the user's account
...
    VoiceState:
      type: object
      properties:
        guildId:
          type: string
          description: Guild ID this voice state is for
        channelId:
          type: string
          description: Channel ID this user is connected to
        userId:
          type: string
          description: User ID this voice state is for
        member:
          $ref: '#/components/schemas/GuildMember'
          description: Guild member this voice state is for
        sessionId:
          type: string
          description: Session ID for this voice state
        deaf:
          type: boolean
          description: Whether this user is deafened by the server
        mute:
          type: boolean
          description: Whether this user is muted by the server
        selfDeaf:
          type: boolean
          description: Whether this user is locally deafened
        selfMute:
          type: boolean
          description: Whether this user is locally muted
        selfStream:
          type: boolean
          description: Whether this user is streaming using "Go Live"
        selfVideo:
          type: boolean
          description: Whether this user's camera is enabled
        suppress:
          type: boolean
          description: Whether this user is muted by the current user
        requestToSpeakTimestamp:
          type: string
          format: date-time
          description: Time at which the user requested to speak

    # Role related schemas
    Role:
      type: object
      required:
        - id
        - name
        - color
        - hoist
        - position
        - permissions
        - managed
        - mentionable
      properties:
        id:
          type: string
          description: Role ID
        name:
          type: string
          description: Role name
        color:
          type: integer
          description: Integer representation of hexadecimal color code
        hoist:
          type: boolean
          description: Whether this role is pinned in the user listing
        icon:
          type: string
          description: Role icon hash
        unicodeEmoji:
          type: string
          description: Role unicode emoji
        position:
          type: integer
          description: Position of this role
        permissions:
          type: string
          description: Permission bit set
        managed:
          type: boolean
          description: Whether this role is managed by an integration
        mentionable:
          type: boolean
          description: Whether this role can be mentioned
        tags:
          $ref: '#/components/schemas/RoleTags'
          description: Role tags object
...
    Emoji:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Emoji ID
        name:
          type: string
          description: Emoji name
        roles:
          type: array
          items:
            type: string
          description: Roles allowed to use this emoji
        user:
          $ref: '#/components/schemas/User'
          description: User that created this emoji
        requireColons:
          type: boolean
          description: Whether this emoji must be wrapped in colons
        managed:
          type: boolean
          description: Whether this emoji is managed
        animated:
          type: boolean
          description: Whether this emoji is animated
        available:
          type: boolean
          description: Whether this emoji can be used

    # Sticker related schemas
...